--- src/api/domain/GuiModel.ts
@@ -1,8 +1,0 @@
-import GuiModelMetadata from "./GuiModelMetadata.ts";
-
-export type GuiModel = {
-    request?: any[];
-    response?: any[];
-    metadata?: GuiModelMetadata;
-    fromFile?: boolean;
-}
